+++ CHANGELOG.md
@@ -2,11 +2,14 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+_(none)_
+
+--------------------
+
+## 4.0.0 (2016-09-12)
 * read and fetch as seperate options for Loader
 * Add seperate JSON and JSON5 types
 * Improved error handling
-
---------------------
 
 ## 3.1.0 (2016-09-06)
 _(none)_